Create main.cpp
This code defines a Ticket class and a Bus class. The Ticket class is used to create objects that represent tickets for a bus. The Bus class is used to create objects that represent buses, each of which has a unique number and a list of stops.

When an instance of the Ticket class is created, a ticket number is generated using the current system time and the rest of the ticket data is passed as arguments to the constructor function. The Ticket class also has methods that allow the user to print the ticket and write its data to a file called tickets.csv.

The Bus class has a constructor that takes a bus number and a list of stops as arguments and sets the corresponding fields in the object. The Bus class also has methods that allow the user to retrieve the bus number, stops, and the number of tickets that have been sold for the bus. It also has a method called isFull that checks whether the maximum number of tickets for the bus has been reached. The maximum number of tickets is set to 10 by default, but this can be changed by modifying the MAX_TICKETS constant at the top of the code.

The code also defines a main function that allows the user to create instances of the Ticket and Bus classes and use their methods. The main function starts by creating a vector called buses that contains three Bus objects with different bus numbers and stop lists. It then prompts the user to enter their name and the bus number they want to travel on. If the bus number is valid, the code creates a Ticket object for the user and writes its data to the tickets.csv file. It then prints the ticket and updates the number of tickets sold for the selected bus in the buses vector. If the bus is full, the code displays an error message and does not create a ticket.